requirements.txt
python-dotenv
numpy
matplotlib